requirements.txt
fpdf==1.7.2
SQLAlchemy==1.3.17
python-telegram-bot==12.7
requests~=2.23.0
